requirements.txt
numpy
requests
matplotlib
torch
whatimage
streamlit
Pillow
opencv-python
pyaml
jupyter
